Tweaks for browser compatibility mostly...
Also added option for secure connections back in, but this is only
allowed in Firefox now because of issues wither other browsers...

Javascript for adding new elements to the school form was updated to
work on all browsers as well.

Tested on IE 7, Chrome, Firefox 3, and Safari 4. If anyone uses Opera,
well oh well.